Move team/thread sort to Experimental::, new header
