+++ algorithms/src/Kokkos_Sort.hpp
@@ -50,10 +50,8 @@
 #endif
 
 #include <Kokkos_Core.hpp>
-#include <std_algorithms/impl/Kokkos_HelperPredicates.hpp>
-#include <std_algorithms/Kokkos_Swap.hpp>
-
 #include <algorithm>
+#include <Kokkos_NestedSort.hpp>
 
 namespace Kokkos {
 
@@ -665,181 +663,6 @@ void sort(ViewType view, size_t const begin, size_t const end) {
   typename ViewType::execution_space exec;
   sort(exec, view, begin, end);
   exec.fence("Kokkos::Sort: fence after sorting");
-}
-
-namespace Impl {
-
-// true for TeamVectorRange, false for ThreadVectorRange
-template <bool teamLevel>
-struct NestedRange {};
-
-// Specialization for team-level
-template <>
-struct NestedRange<true> {
-  template <typename TeamMember, typename SizeType>
-  KOKKOS_FUNCTION static auto create(const TeamMember& t, SizeType len) {
-    return Kokkos::TeamVectorRange(t, len);
-  }
-  template <typename TeamMember>
-  KOKKOS_FUNCTION static void barrier(const TeamMember& t) {
-    t.team_barrier();
-  }
-};
-
-// Specialization for thread-level
-template <>
-struct NestedRange<false> {
-  template <typename TeamMember, typename SizeType>
-  KOKKOS_FUNCTION static auto create(const TeamMember& t, SizeType len) {
-    return Kokkos::ThreadVectorRange(t, len);
-  }
-  // Barrier is no-op, as vector lanes of a thread are implicitly synchronized
-  // after parallel region
-  template <typename TeamMember>
-  KOKKOS_FUNCTION static void barrier(const TeamMember&) {}
-};
-
-// When just doing sort (not sort_by_key), use nullptr_t for ValueViewType.
-// This only takes the NestedRange instance for template arg deduction.
-template <class TeamMember, class KeyViewType, class ValueViewType,
-          class Comparator, bool useTeamLevel>
-KOKKOS_INLINE_FUNCTION void sort_nested_impl(const TeamMember& t,
-                                             const KeyViewType& keyView,
-                                             const ValueViewType& valueView,
-                                             const Comparator& comp,
-                                             const NestedRange<useTeamLevel>&) {
-  using SizeType  = typename KeyViewType::size_type;
-  using KeyType   = typename KeyViewType::non_const_value_type;
-  using Range     = NestedRange<useTeamLevel>;
-  SizeType n      = keyView.extent(0);
-  SizeType npot   = 1;
-  SizeType levels = 0;
-  // FIXME: ceiling power-of-two is a common thing to need - make it a utility
-  while (npot < n) {
-    levels++;
-    npot <<= 1;
-  }
-  for (SizeType i = 0; i < levels; i++) {
-    for (SizeType j = 0; j <= i; j++) {
-      // n/2 pairs of items are compared in parallel
-      Kokkos::parallel_for(Range::create(t, npot / 2), [=](const SizeType k) {
-        // How big are the brown/pink boxes?
-        // (Terminology comes from Wikipedia diagram)
-        // https://commons.wikimedia.org/wiki/File:BitonicSort.svg#/media/File:BitonicSort.svg
-        SizeType boxSize = SizeType(2) << (i - j);
-        // Which box contains this thread?
-        SizeType boxID     = k >> (i - j);          // k * 2 / boxSize;
-        SizeType boxStart  = boxID << (1 + i - j);  // boxID * boxSize
-        SizeType boxOffset = k - (boxStart >> 1);   // k - boxID * boxSize / 2;
-        SizeType elem1     = boxStart + boxOffset;
-        if (j == 0) {
-          // first phase (brown box): within a block, compare with the
-          // opposite value in the box
-          SizeType elem2 = boxStart + boxSize - 1 - boxOffset;
-          if (elem2 < n) {
-            KeyType key1 = keyView(elem1);
-            KeyType key2 = keyView(elem2);
-            if (comp(key2, key1)) {
-              keyView(elem1) = key2;
-              keyView(elem2) = key1;
-              if constexpr (!std::is_same_v<ValueViewType, std::nullptr_t>) {
-                Kokkos::Experimental::swap(valueView(elem1), valueView(elem2));
-              }
-            }
-          }
-        } else {
-          // later phases (pink box): within a block, compare with fixed
-          // distance (boxSize / 2) apart
-          SizeType elem2 = elem1 + boxSize / 2;
-          if (elem2 < n) {
-            KeyType key1 = keyView(elem1);
-            KeyType key2 = keyView(elem2);
-            if (comp(key2, key1)) {
-              keyView(elem1) = key2;
-              keyView(elem2) = key1;
-              if constexpr (!std::is_same_v<ValueViewType, std::nullptr_t>) {
-                Kokkos::Experimental::swap(valueView(elem1), valueView(elem2));
-              }
-            }
-          }
-        }
-      });
-      Range::barrier(t);
-    }
-  }
-}
-
-}  // namespace Impl
-
-template <class TeamMember, class ViewType>
-KOKKOS_INLINE_FUNCTION void sort_team(const TeamMember& t,
-                                      const ViewType& view) {
-  Impl::sort_nested_impl(t, view, nullptr,
-                         Experimental::Impl::StdAlgoLessThanBinaryPredicate<
-                             typename ViewType::non_const_value_type>(),
-                         Impl::NestedRange<true>());
-}
-
-template <class TeamMember, class ViewType, class Comparator>
-KOKKOS_INLINE_FUNCTION void sort_team(const TeamMember& t, const ViewType& view,
-                                      const Comparator& comp) {
-  Impl::sort_nested_impl(t, view, nullptr, comp, Impl::NestedRange<true>());
-}
-
-template <class TeamMember, class KeyViewType, class ValueViewType>
-KOKKOS_INLINE_FUNCTION void sort_by_key_team(const TeamMember& t,
-                                             const KeyViewType& keyView,
-                                             const ValueViewType& valueView) {
-  Impl::sort_nested_impl(t, keyView, valueView,
-                         Experimental::Impl::StdAlgoLessThanBinaryPredicate<
-                             typename KeyViewType::non_const_value_type>(),
-                         Impl::NestedRange<true>());
-}
-
-template <class TeamMember, class KeyViewType, class ValueViewType,
-          class Comparator>
-KOKKOS_INLINE_FUNCTION void sort_by_key_team(const TeamMember& t,
-                                             const KeyViewType& keyView,
-                                             const ValueViewType& valueView,
-                                             const Comparator& comp) {
-  Impl::sort_nested_impl(t, keyView, valueView, comp,
-                         Impl::NestedRange<true>());
-}
-
-template <class TeamMember, class ViewType>
-KOKKOS_INLINE_FUNCTION void sort_thread(const TeamMember& t,
-                                        const ViewType& view) {
-  Impl::sort_nested_impl(t, view, nullptr,
-                         Experimental::Impl::StdAlgoLessThanBinaryPredicate<
-                             typename ViewType::non_const_value_type>(),
-                         Impl::NestedRange<false>());
-}
-
-template <class TeamMember, class ViewType, class Comparator>
-KOKKOS_INLINE_FUNCTION void sort_thread(const TeamMember& t,
-                                        const ViewType& view,
-                                        const Comparator& comp) {
-  Impl::sort_nested_impl(t, view, nullptr, comp, Impl::NestedRange<false>());
-}
-
-template <class TeamMember, class KeyViewType, class ValueViewType>
-KOKKOS_INLINE_FUNCTION void sort_by_key_thread(const TeamMember& t,
-                                               const KeyViewType& keyView,
-                                               const ValueViewType& valueView) {
-  Impl::sort_nested_impl(t, keyView, valueView,
-                         Experimental::Impl::StdAlgoLessThanBinaryPredicate<
-                             typename KeyViewType::non_const_value_type>(),
-                         Impl::NestedRange<false>());
-}
-
-template <class TeamMember, class KeyViewType, class ValueViewType,
-          class Comparator>
-KOKKOS_INLINE_FUNCTION void sort_by_key_thread(const TeamMember& t,
-                                               const KeyViewType& keyView,
-                                               const ValueViewType& valueView,
-                                               const Comparator& comp) {
-  Impl::sort_nested_impl(t, keyView, valueView, comp,
-                         Impl::NestedRange<false>());
 }
 
 }  // namespace Kokkos
